Move DocumentValidationRulesFactory.cs out of ChargeCommand folder as its now being used by multiple commnads
