+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/Dtos/Validation/DocumentValidation/ValidationRules/BusinessReasonCodeMustBeUpdateChargeInformationOrChargePricesRule.cs
@@ -13,22 +13,23 @@
 // limitations under the License.
 
 using GreenEnergyHub.Charges.Domain.Dtos.SharedDtos;
-using GreenEnergyHub.Charges.Domain.Dtos.Validation;
 
-namespace GreenEnergyHub.Charges.Domain.Dtos.ChargeInformationCommands.Validation.DocumentValidation.ValidationRules
+namespace GreenEnergyHub.Charges.Domain.Dtos.Validation.DocumentValidation.ValidationRules
 {
-    public class SenderIsMandatoryTypeValidationRule : IValidationRule
+    public class BusinessReasonCodeMustBeUpdateChargeInformationOrChargePricesRule : IValidationRule
     {
         private readonly DocumentDto _documentDto;
 
-        public SenderIsMandatoryTypeValidationRule(DocumentDto documentDto)
+        public BusinessReasonCodeMustBeUpdateChargeInformationOrChargePricesRule(DocumentDto documentDto)
         {
             _documentDto = documentDto;
         }
 
         public ValidationRuleIdentifier ValidationRuleIdentifier =>
-            ValidationRuleIdentifier.SenderIsMandatoryTypeValidation;
+            ValidationRuleIdentifier.BusinessReasonCodeMustBeUpdateChargeInformationOrChargePrices;
 
-        public bool IsValid => MarketParticipantMrIdValidator.IsValid(_documentDto.Sender.MarketParticipantId);
+        public bool IsValid =>
+            _documentDto.BusinessReasonCode is BusinessReasonCode.UpdateChargeInformation
+                or BusinessReasonCode.UpdateChargePrices;
     }
 }
